remove abstract render context type.
