--- SoftGL/RenderContext/GLRenderContext.IDisposable.cs
@@ -1,50 +1,0 @@
-﻿using System;
-
-namespace SoftGL
-{
-    /// <summary>
-    ///
-    /// </summary>
-    abstract partial class GLRenderContext
-    {
-        /// <summary>
-        ///
-        /// </summary>
-        public void Dispose()
-        {
-            this.Dispose(true);
-            GC.SuppressFinalize(this);
-        }
-
-        /// <summary>
-        ///
-        /// </summary>
-        ~GLRenderContext()
-        {
-            this.Dispose(false);
-        }
-
-        private bool disposedValue = false;
-
-        private void Dispose(bool disposing)
-        {
-            if (this.disposedValue == false)
-            {
-                if (disposing)
-                {
-                    // Dispose managed resources.
-                }
-
-                // Dispose unmanaged resources.
-                DisposeUnmanagedResources();
-            }
-
-            this.disposedValue = true;
-        }
-
-        /// <summary>
-        ///
-        /// </summary>
-        protected abstract void DisposeUnmanagedResources();
-    }
-}
